Handle event passing between the board and the UI
When a cell is opened this can cause ther cells to be opened as well. In
order to achieve this we need the state of all the cells to be managed
at a higher place in the hierarchy, so it can be passed down to them.

Additionally now we need to treat the `Game` object itself as state for
the `Board` component, since we need to keep it consistent between
rerenders, which will happen when the cell states change.